Update peer
1. Update Perr.ServeConn
2. Change the type of Perr.defaultReadTimeout and peer.defaultWriteTimeout to ‘time.Duration’

+++ session.go
@@ -149,8 +149,8 @@ func newSession(peer *Peer, conn net.Conn, protoFuncs []socket.ProtoFunc) *sessi
 		pushRouter:   peer.PushRouter,
 		socket:       socket.NewSocket(conn, protoFuncs...),
 		pullCmdMap:   goutil.RwMap(),
-		readTimeout:  peer.defaultReadTimeout,
-		writeTimeout: peer.defaultWriteTimeout,
+		readTimeout:  int32(peer.defaultReadTimeout),
+		writeTimeout: int32(peer.defaultWriteTimeout),
 	}
 	return s
 }
